feat: Complete project reorganization with Pydantic models
- Split personality.py into personalities/ directory:
  - base.py: Abstract base class LLMPersonality
  - claude.py: Claude implementation
  - openai.py: OpenAI implementation
  - local.py: Local model implementation
  - factory.py: create_personality factory function

- Update all imports to use new structure
- Add backward compatibility in personality.py
- Extend PersonalityConfig with all needed fields for compatibility
- Fix relative imports to use absolute imports
- Update debate_app.py to use Pydantic models

The codebase is now more modular with clear separation of concerns:
- models/: All Pydantic data models with validation
- services/: Business logic (debate management, file handling)
- ui/: User interface components
- personalities/: AI personality implementations
