+++ personalities/__init__.py
@@ -1,0 +1,15 @@
+"""AI personality implementations for the ASS debate system."""
+
+from .base import LLMPersonality
+from .claude import ClaudePersonality
+from .openai import OpenAIPersonality
+from .local import LocalModelPersonality
+from .factory import create_personality
+
+__all__ = [
+    "LLMPersonality",
+    "ClaudePersonality", 
+    "OpenAIPersonality",
+    "LocalModelPersonality",
+    "create_personality",
+]
